+++ lib/ept_sdk/entity.ex
@@ -8,7 +8,8 @@ defmodule EPTSDK.Entity do
           "type" => "payment_methods" = type,
           "attributes" => attributes
         } = record,
-        links
+        links,
+        included
       ),
       do: %EPTSDK.PaymentMethod{
         id: id,
@@ -24,6 +25,11 @@ defmodule EPTSDK.Entity do
         discarded_at: fetch_datetime(attributes, "discarded_at"),
         created_at: fetch_datetime(attributes, "created_at"),
         updated_at: fetch_datetime(attributes, "updated_at"),
+        merchant_account:
+          fetch_relationship(record["relationships"], "merchant_account", included),
+        charges: fetch_relationship(record["relationships"], "charges", included),
+        customer: fetch_relationship(record["relationships"], "customer", included),
+        address: fetch_relationship(record["relationships"], "address", included),
         __relationships__: record["relationships"],
         __links__: record["links"] || links,
         __raw__: record
@@ -35,16 +41,34 @@ defmodule EPTSDK.Entity do
           "type" => "charges" = type,
           "attributes" => attributes
         } = record,
-        links
+        links,
+        included
       ),
       do: %EPTSDK.Charge{
         id: id,
         type: type,
+        amount: fetch_money(attributes, ["amount_cents", "currency"]),
         amount_cents: fetch(attributes, "amount_cents"),
+        fee: fetch_money(attributes, ["fee_cents", "currency"]),
+        fee_cents: fetch(attributes, "fee_cents"),
+        gateway: fetch_money(attributes, ["gateway_cents", "currency"]),
+        gateway_cents: fetch(attributes, "gateway_cents"),
+        refund: fetch_money(attributes, ["refund_cents", "currency"]),
+        refund_cents: fetch(attributes, "refund_cents"),
+        net: fetch_money(attributes, ["net_cents", "currency"]),
+        net_cents: fetch(attributes, "net_cents"),
         currency: fetch(attributes, "currency"),
         description: fetch(attributes, "description"),
         created_at: fetch_datetime(attributes, "created_at"),
         updated_at: fetch_datetime(attributes, "updated_at"),
+        customer: fetch_relationship(record["relationships"], "customer", included),
+        payment_method: fetch_relationship(record["relationships"], "payment_method", included),
+        billing_address: fetch_relationship(record["relationships"], "billing_address", included),
+        subscription: fetch_relationship(record["relationships"], "subscription", included),
+        shipping_address:
+          fetch_relationship(record["relationships"], "shipping_address", included),
+        merchant_account:
+          fetch_relationship(record["relationships"], "merchant_account", included),
         __relationships__: record["relationships"],
         __links__: record["links"] || links,
         __raw__: record
@@ -56,9 +80,10 @@ defmodule EPTSDK.Entity do
           "type" => "consumer_addresses" = type,
           "attributes" => attributes
         } = record,
-        links
+        links,
+        included
       ),
-      do: %EPTSDK.Address{
+      do: %EPTSDK.ConsumerAddress{
         id: id,
         type: type,
         line_1: fetch(attributes, "line_1"),
@@ -69,6 +94,8 @@ defmodule EPTSDK.Entity do
         country: fetch(attributes, "country"),
         created_at: fetch_datetime(attributes, "created_at"),
         updated_at: fetch_datetime(attributes, "updated_at"),
+        merchant_account:
+          fetch_relationship(record["relationships"], "merchant_account", included),
         __relationships__: record["relationships"],
         __links__: record["links"] || links,
         __raw__: record
@@ -80,7 +107,8 @@ defmodule EPTSDK.Entity do
           "type" => "customers" = type,
           "attributes" => attributes
         } = record,
-        links
+        links,
+        included
       ),
       do: %EPTSDK.Customer{
         id: id,
@@ -92,6 +120,11 @@ defmodule EPTSDK.Entity do
         blocked_at: fetch_datetime(attributes, "blocked_at"),
         created_at: fetch_datetime(attributes, "created_at"),
         updated_at: fetch_datetime(attributes, "updated_at"),
+        address: fetch_relationship(record["relationships"], "address", included),
+        payment_methods: fetch_relationship(record["relationships"], "payment_methods", included),
+        merchant_account:
+          fetch_relationship(record["relationships"], "merchant_account", included),
+        charges: fetch_relationship(record["relationships"], "charges", included),
         # TODO: turn into formal relationship structs
         __relationships__: record["relationships"],
         # TODO: turn into formal links structs
@@ -105,7 +138,8 @@ defmodule EPTSDK.Entity do
           "type" => "merchant_accounts" = type,
           "attributes" => attributes
         } = record,
-        links
+        links,
+        included
       ),
       do: %EPTSDK.MerchantAccount{
         id: id,
@@ -137,6 +171,17 @@ defmodule EPTSDK.Entity do
         pending_at: fetch_datetime(attributes, "pending_at"),
         created_at: fetch_datetime(attributes, "created_at"),
         updated_at: fetch_datetime(attributes, "updated_at"),
+        customers: fetch_relationship(record["relationships"], "customers", included),
+        consumer_addresses:
+          fetch_relationship(record["relationships"], "consumer_addresses", included),
+        payment_methods: fetch_relationship(record["relationships"], "payment_methods", included),
+        charges: fetch_relationship(record["relationships"], "charges", included),
+        payouts: fetch_relationship(record["relationships"], "payouts", included),
+        payout_method: fetch_relationship(record["relationships"], "payout_method", included),
+        events: fetch_relationship(record["relationships"], "events", included),
+        webhook_subscriptions:
+          fetch_relationship(record["relationships"], "webhook_subscriptions", included),
+        subscriptions: fetch_relationship(record["relationships"], "subscriptions", included),
         # TODO: turn into formal relationship structs
         __relationships__: record["relationships"],
         # TODO: turn into formal links structs
@@ -150,7 +195,8 @@ defmodule EPTSDK.Entity do
           "type" => "events" = type,
           "attributes" => attributes
         } = record,
-        links
+        links,
+        included
       ),
       do: %EPTSDK.Event{
         id: id,
@@ -159,26 +205,8 @@ defmodule EPTSDK.Entity do
         payload: fetch(attributes, "payload"),
         occurred_at: fetch_datetime(attributes, "occurred_at"),
         created_at: fetch_datetime(attributes, "created_at"),
-        # TODO: turn into formal relationship structs
-        __relationships__: record["relationships"],
-        # TODO: turn into formal links structs
-        __links__: record["links"] || links,
-        __raw__: record
-      }
-
-  def to_struct(
-        %{
-          "id" => id,
-          "type" => "disputes" = type,
-          "attributes" => attributes
-        } = record,
-        links
-      ),
-      do: %EPTSDK.Dispute{
-        id: id,
-        type: type,
-        created_at: fetch_datetime(attributes, "created_at"),
-        updated_at: fetch_datetime(attributes, "updated_at"),
+        merchant_account:
+          fetch_relationship(record["relationships"], "merchant_account", included),
         # TODO: turn into formal relationship structs
         __relationships__: record["relationships"],
         # TODO: turn into formal links structs
@@ -192,7 +220,8 @@ defmodule EPTSDK.Entity do
           "type" => "payout_methods" = type,
           "attributes" => attributes
         } = record,
-        links
+        links,
+        included
       ),
       do: %EPTSDK.PayoutMethod{
         id: id,
@@ -205,6 +234,9 @@ defmodule EPTSDK.Entity do
         verified_at: fetch_datetime(attributes, "verified_at"),
         created_at: fetch_datetime(attributes, "created_at"),
         updated_at: fetch_datetime(attributes, "updated_at"),
+        merchant_account:
+          fetch_relationship(record["relationships"], "merchant_account", included),
+        payouts: fetch_relationship(record["relationships"], "payouts", included),
         # TODO: turn into formal relationship structs
         __relationships__: record["relationships"],
         # TODO: turn into formal links structs
@@ -218,7 +250,8 @@ defmodule EPTSDK.Entity do
           "type" => "subscriptions" = type,
           "attributes" => attributes
         } = record,
-        links
+        links,
+        included
       ),
       do: %EPTSDK.Subscription{
         id: id,
@@ -234,6 +267,11 @@ defmodule EPTSDK.Entity do
         end_at: fetch_datetime(attributes, "end_at"),
         created_at: fetch_datetime(attributes, "created_at"),
         updated_at: fetch_datetime(attributes, "updated_at"),
+        merchant_account:
+          fetch_relationship(record["relationships"], "merchant_account", included),
+        charges: fetch_relationship(record["relationships"], "charges", included),
+        payment_method: fetch_relationship(record["relationships"], "payment_method", included),
+        customer: fetch_relationship(record["relationships"], "customer", included),
         # TODO: turn into formal relationship structs
         __relationships__: record["relationships"],
         # TODO: turn into formal links structs
@@ -247,7 +285,8 @@ defmodule EPTSDK.Entity do
           "type" => "webhook_subscriptions" = type,
           "attributes" => attributes
         } = record,
-        links
+        links,
+        included
       ),
       do: %EPTSDK.WebhookSubscription{
         id: id,
@@ -260,6 +299,8 @@ defmodule EPTSDK.Entity do
         url: fetch(attributes, "url"),
         created_at: fetch_datetime(attributes, "created_at"),
         updated_at: fetch_datetime(attributes, "updated_at"),
+        merchant_account:
+          fetch_relationship(record["relationships"], "merchant_account", included),
         # TODO: turn into formal relationship structs
         __relationships__: record["relationships"],
         # TODO: turn into formal links structs
@@ -273,7 +314,8 @@ defmodule EPTSDK.Entity do
           "type" => "payouts" = type,
           "attributes" => attributes
         } = record,
-        links
+        links,
+        included
       ),
       do: %EPTSDK.Payout{
         id: id,
@@ -288,6 +330,9 @@ defmodule EPTSDK.Entity do
         transfer_credit_type: fetch(attributes, "transfer_credit_type"),
         created_at: fetch_datetime(attributes, "created_at"),
         updated_at: fetch_datetime(attributes, "updated_at"),
+        merchant_account:
+          fetch_relationship(record["relationships"], "merchant_account", included),
+        payout_method: fetch_relationship(record["relationships"], "payout_method", included),
         # TODO: turn into formal relationship structs
         __relationships__: record["relationships"],
         # TODO: turn into formal links structs
@@ -295,27 +340,98 @@ defmodule EPTSDK.Entity do
         __raw__: record
       }
 
-  def to_struct(record, links) when not is_map_key(record, "attributes") do
-    to_struct(Map.put(record, "attributes", %{}), links)
+  def to_struct(record, links, included) when not is_map_key(record, "attributes") do
+    to_struct(Map.put(record, "attributes", %{}), links, included)
   end
 
   defp fetch(attributes, key) when is_map(attributes) and is_binary(key) do
     if Map.has_key?(attributes, key) do
       attributes[key]
     else
-      %EPTSDK.PropertyNotAvailable{property: key, reason: :unfetched}
+      %EPTSDK.PropertyNotAvailable{name: key, reason: :unfetched}
+    end
+  end
+
+  defp fetch_money(attributes, [cents_key, currency_key])
+       when is_map(attributes) and is_binary(cents_key) and is_binary(currency_key) do
+    with true <- Map.has_key?(attributes, cents_key),
+         true <- Map.has_key?(attributes, currency_key),
+         cents when is_integer(cents) <- Map.get(attributes, cents_key),
+         currency when is_binary(currency) <- Map.get(attributes, currency_key) do
+      Money.new(cents, currency)
+    else
+      false -> %EPTSDK.PropertyNotAvailable{name: [cents_key, currency_key], reason: :unfetched}
+      nil -> nil
+      _ -> %EPTSDK.PropertyNotAvailable{name: [cents_key, currency_key], reason: :decoding_error}
     end
   end
 
   defp fetch_datetime(attributes, key) when is_map(attributes) and is_binary(key) do
     with true <- Map.has_key?(attributes, key),
-         value <- Map.get(attributes, key),
+         value when not is_nil(value) <- Map.get(attributes, key),
          {:ok, timestamp} <- Timex.parse(value, "{ISO:Extended:Z}") do
       timestamp
     else
+      false -> %EPTSDK.PropertyNotAvailable{name: key, reason: :unfetched}
       nil -> nil
-      false -> %EPTSDK.PropertyNotAvailable{property: key, reason: :unfetched}
-      {:error, _message} -> %EPTSDK.PropertyNotAvailable{property: key, reason: :decoding_error}
+      {:error, _message} -> %EPTSDK.PropertyNotAvailable{name: key, reason: :decoding_error}
     end
   end
+
+  defp fetch_relationship(relationships, key, []) do
+    Map.get(relationships, key)
+    |> case do
+      nil ->
+        %EPTSDK.RelationshipNotAvailable{
+          name: key,
+          reason: :undefined_relationship
+        }
+
+      _relationship ->
+        %EPTSDK.RelationshipNotAvailable{name: key, reason: :not_included}
+    end
+  end
+
+  defp fetch_relationship(relationships, key, included)
+       when is_map(relationships) and is_binary(key) and is_list(included) do
+    Map.get(relationships, key)
+    |> case do
+      nil -> {:error, :unknown_relationship}
+      relationship -> {:ok, decode_relationship(relationship)}
+    end
+    |> case do
+      {:error, _} = error -> error
+      {:ok, nil} -> {:ok, []}
+      {:ok, relation} -> {:ok, find_by_relation(included, relation)}
+    end
+    |> case do
+      {:error, _} = error -> error
+      {:ok, nil} -> {:error, :not_included}
+      {:ok, data} when is_list(data) -> data |> Enum.map(&to_struct(&1, nil, included))
+      {:ok, data} -> to_struct(data, nil, included)
+    end
+    |> case do
+      {:error, reason} ->
+        %EPTSDK.RelationshipNotAvailable{
+          name: key,
+          reason: reason
+        }
+
+      record ->
+        record
+    end
+  end
+
+  defp decode_relationship(%{"data" => %{"id" => id, "type" => type}} = relationship)
+       when is_map(relationship) do
+    {id, type}
+  end
+
+  defp decode_relationship(%{} = relationship) when is_map(relationship), do: nil
+
+  defp find_by_relation(included, {id, type}),
+    do:
+      Enum.find(included, fn %{"id" => included_id, "type" => included_type} ->
+        included_type == type && included_id == id
+      end)
 end