+++ lib/ept_sdk/relationship_not_available.ex
@@ -1,4 +1,4 @@
-defmodule EPTSDK.PropertyNotAvailable do
+defmodule EPTSDK.RelationshipNotAvailable do
   @enforce_keys [:name, :reason]
   defstruct [:name, :reason]
   # TODO: What if we could turn this into a request to fill the data?