+++ lib/ept_sdk/webhook_subscription.ex
@@ -7,6 +7,7 @@ defmodule EPTSDK.WebhookSubscription do
     :id,
     :created_at,
     :updated_at,
+    :merchant_account,
     :__raw__,
     :__links__,
     :__relationships__
@@ -22,6 +23,7 @@ defmodule EPTSDK.WebhookSubscription do
     :url,
     :created_at,
     :updated_at,
+    :merchant_account,
     :__raw__,
     :__links__,
     :__relationships__